+++ src/core/layers/layer.js
@@ -54,6 +54,7 @@ export default function Layer (attrs, layer, initialize = true) {
     else if (attrs.isSelected === true || store.getters.treeType === "light") {
         this.setIsVisibleInMap(attrs.isSelected);
     }
+    this.setMinMaxResolutions();
     this.checkForScale({scale: store.getters["Map/scale"]});
     this.registerInteractionMapViewListeners();
 }
@@ -498,6 +499,17 @@ Layer.prototype.showLayerInformation = function () {
 Layer.prototype.setLegend = function (value) {
     this.set("legend", value);
     store.dispatch("Legend/setLegendOnChanged", value);
+};
+/**
+ * Sets visible min and max resolution on layer.
+ * @returns {void}
+ */
+Layer.prototype.setMinMaxResolutions = function () {
+    const resoByMaxScale = bridge.getResoByScale(this.get("maxScale"), "max"),
+        resoByMinScale = bridge.getResoByScale(this.get("minScale"), "min");
+
+    this.get("layer").setMaxResolution(resoByMaxScale + (resoByMaxScale / 100));
+    this.get("layer").setMinResolution(resoByMinScale);
 };
 
 // backbone-relevant functions (may be removed if all layers are no longer backbone models):