+++ CHANGELOG.md
@@ -51,6 +51,7 @@
 - Searchbar: Searches that deliver only a single search result can now be selected with `Enter`.
 - Handling of Multipolygons in WKT geometry parsing
 - GFI: in the iframe the content is now displayed again when using the `desktopType` `attached`.
+- GFI for theme datatable is displayed. Resolution at wms layer is set correctly.
 
 
 ---